add id field directly in heading if possible
Instead of generating a useless <span id="..."> before the relevant
<h?> tag, just add the id field to the <h?> tag if it does not have
one already.

This way, the target of the link is exactly the <h?> tag, and the
browser positions the focus exactly on it. Without this, there was a
slight difference in offset between tags which had an id= field an
tags which didn't.